ng update and interceptor as a function
