+++ package.json
@@ -12,7 +12,7 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^18.0.1",
-    "@angular/cdk": "^17.3.9",
+    "@angular/cdk": "^18.0.1",
     "@angular/common": "^18.0.1",
     "@angular/compiler": "^18.0.1",
     "@angular/core": "^18.0.1",
@@ -41,11 +41,11 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^18.0.2",
-    "@angular-eslint/builder": "17.4.1",
-    "@angular-eslint/eslint-plugin": "17.4.1",
-    "@angular-eslint/eslint-plugin-template": "17.4.1",
-    "@angular-eslint/schematics": "17.4.1",
-    "@angular-eslint/template-parser": "17.4.1",
+    "@angular-eslint/builder": "18.0.1",
+    "@angular-eslint/eslint-plugin": "18.0.1",
+    "@angular-eslint/eslint-plugin-template": "18.0.1",
+    "@angular-eslint/schematics": "18.0.1",
+    "@angular-eslint/template-parser": "18.0.1",
     "@angular/cli": "^18.0.2",
     "@angular/compiler-cli": "^18.0.1",
     "@types/clownface": "^2.0.7",
