+++ projects/blueprint/src/app/core/http-interceptor/index.ts
@@ -1,1 +1,1 @@
-export * from './auth-interceptor/auth-interceptor.service'
+export * from './auth-interceptor/auth-interceptor.function'
