+++ projects/blueprint/src/index.html
@@ -25,7 +25,7 @@
 </head>
 
 <body>
-  <bp-root />
+  <bp-root></bp-root>
   <noscript>Please enable JavaScript to continue using this application.</noscript>
 </body>
